E2E tests: wire up the VRF logic
We add a new "VRF" field to the external containers, which is the VRF
the container is supposed to be reached through. We then change all
the places where we check the session with the ip of the node, taking
the ip of the interface inside the vrf instead of the node's public
IP.

Finally, we add a simple vrf test that validates the two containers
work.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>