Wipe out the loadbalancers assigned to a service if they are not valid
Currently, when metallb receives a service with malformed lb ips and
it's not able to determine the ip family, it errors out and reprocesses
the service.

Here if the lb part of the status is invalid we just wipe it and let the
assignement logic add a new ip. This prevents malformed services to
cause infinite loops and helps in recovering the service.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>