e2etest: remove the forced ips for the containers
Passing the ips from outside never worked, because the ips were passed
after the image's name and so were injected as parameters of the entry
point. Moreover, even if that was fixed, that won't work with some
docker version resulting in a "user specified IP address is supported
only when connecting to networks with user configured subnets" error.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>