FRR Mode: provide a way to append a piece of configuration
Currently it is not possible to run multiple FRR instances on the same
node, due to port clashing in first instance.
In order to provide a way to overcome this issue, we provide a non
supported - use at your own risk way to append a fixed configuration
right after the one generated by metallb. This can be used in short term
to mitigate some of the issues reported.

This solution will be deprecated in the long term, being replaced by a
more cohesive one.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>