Fix syncState response for unconverged services
When convergeBalancer fails to assign an LB IP to a
unallocated service, the response of the service handler
was SyncStateSuccess which errounously indicates that the service
converged. This PR sets the response to SyncStateErrorNoRetry.

No changes in functionaly, just an extra error log message from
the switch of the service controller.

Signed-off-by: Marcelo Guerrero Viveros <marguerr@redhat.com>