fix: add missing namespace to helm chart templates
Helm template, unlike helm install, does not set the namespace for namespaced resources.
This inconsistency creates issues for users who use helm template to generate the manifests, and then use other tools to apply the manifests (kubectl, terraform, argo etc)
A workaround is to add the namespace field to the metadata section.

Signed-off-by: ankitm123 <ankitmohapatra123@gmail.com>